Семинар 8

+++ php-fpm-port/code/src/Domain/Controllers/UserController.php
@@ -30,16 +30,14 @@ class UserController extends Controller
         $this->repository = new UserRepository();
     }
 
-
     public function actionIndex(): string
     {
-        $user = new User();
         $currentPage = $_GET['page'] ?? 1;
         $alert = $_GET['alert'] ?? false;
         $message = $_SESSION['alert_message'] ?? "База пуста";
 
         return $this->render->renderPage(
-            'user-index.twig',
+            'layout-user/user-index.twig',
             [
                 'alert_message' => $message,
                 'alert' => $alert,
@@ -55,27 +53,47 @@ class UserController extends Controller
         $user = new User();
         $name = $_POST['name'] ?? '';
         $lastname = $_POST['lastname'] ?? '';
-        $birthday = $_POST['lastname'] ?? '';
+        $birthday = $_POST['birthday'] ?? '';
+        $login = $_POST['login'] ?? '';
+        $password = $_POST['password'] ?? '';
+        $confirm_password = $_POST['confirm_password'] ?? '';
 
-        if ($this->validate->validateRequestData([$name, $lastname, $birthday])) {
-            $user->setParamsFromRequestData($name, $lastname, $birthday);
+        $validData = [
+            'name' => $name,
+            'lastname' => $lastname,
+            'birthday' => $birthday,
+            'login' => $login,
+            'password' => [$password, $confirm_password]];
+
+        if ($this->validate->validateRequestData($validData)) {
+            $user->setParamsFromRequestData($name, $lastname, $birthday, $login, $password);
             $this->repository->saveUserFromStorage($user);
 
             $_SESSION['alert_message'] = "Пользователь добавлен";
             header("Location: /user/index/?alert=true");
             die();
         } else {
+            $logMessage = 'При добавлении пользователя не корректные данные';
+            $logMessage .= " | " . "Попытка вызова адреса " . $_SERVER['REQUEST_URI'];
+            Application::$logger->error($logMessage);
             throw new \Exception("Данные не корректны");
         }
     }
 
     #[NoReturn] public function actionDelete(): void
     {
-        $id = isset($_GET['id']) && is_numeric($_GET['id']) ? (int)$_GET['id'] : 0;
+        $id = $_GET['id'];
+        if ($this->repository->exists($id)) {
+            $_SESSION['alert_message'] = $this->repository->deleteUserFromStorage($id);
+            header("Location: /user/index/?alert=true");
+            die();
 
-        $_SESSION['alert_message'] = $this->repository->deleteUserFromStorage($id);
-        header("Location: /user/index/?alert=true");
-        die();
+        } else {
+            $logMessage = 'При удалении пользователь в базе осутствует';
+            $logMessage .= " | " . "Попытка вызова адреса " . $_SERVER['REQUEST_URI'];
+            Application::$logger->error($logMessage);
+            throw new \Exception("Данный пользователь не найден");
+        }
     }
 
     #[NoReturn] public function actionClear(): void
@@ -89,7 +107,7 @@ class UserController extends Controller
     {
         $_SESSION['alert_message'] = "Пусто";
         return $this->render->renderPage(
-            'user-index.twig',
+            'layout-user/user-index.twig',
             [
                 'users' => $this->repository->searchTodayBirthday()
             ]);
@@ -98,12 +116,12 @@ class UserController extends Controller
     public function actionUpdate(): void
     {
         if ($this->repository->exists($_GET['id'])) {
-
+            $login = $_GET['login'] ?? '';
             $name = $_GET['name'] ?? '';
             $lastname = $_GET['lastname'] ?? '';
             $birthday = $_GET['birthday'] ?? '';
 
-            $arrayData = $this->validate->validateUserData($name, $lastname, $birthday);
+            $arrayData = $this->validate->validateUserData($login, $name, $lastname, $birthday);
             $arrayKey['id_user'] = $_GET['id'];
 
             if ($this->repository->updateData('users', $arrayData, $arrayKey)) {
@@ -113,41 +131,24 @@ class UserController extends Controller
                 die();
 
             } else {
+                $logMessage = 'При изменении пользователя не верные данные';
+                $logMessage .= " | " . "Попытка вызова адреса " . $_SERVER['REQUEST_URI'];
+                Application::$logger->error($logMessage);
                 throw new \Exception("Пользователь не изменен, проверьте данные");
             }
         } else {
+            $logMessage = 'При изменении пользователь в базе осутствует';
+            $logMessage .= " | " . "Попытка вызова адреса " . $_SERVER['REQUEST_URI'];
+            Application::$logger->error($logMessage);
             throw new \Exception("Данный пользователь не найден");
         }
     }
 
-    public function actionHash(): string
-    {
-        if (!empty($_GET['pass_string'])) {
-            return Auth::getPasswordHash($_GET['pass_string']);
-        } else {
-            throw new \Exception("Невозможно сгенерировать хэш. Не передан пароль");
-        }
-
-    }
-
     public function actionAuth(): string
     {
-        $remember = $_COOKIE['remember_token'] ?? '';
-        $user = $this->repository->getAllUserCookie($remember);
-
-        if (!empty($user)) {
-
-            $_SESSION['user_name'] = $user[0]['user_name'];
-            $_SESSION['user_lastname'] = $user[0]['user_lastname'];
-            $_SESSION['id_user'] = $user[0]['id_user'];
-
-            header('Location: /');
-            die();
-        } else {
-            return $this->render->renderPageWithForm([
-                'title' => 'Авторизация',
-            ]);
-        }
+        return $this->render->renderPageWithForm([
+            'title' => 'Авторизация',
+        ]);
     }
 
     public function actionLogin(): string
@@ -155,6 +156,13 @@ class UserController extends Controller
         $result = false;
         if (isset($_POST['login']) && isset($_POST['password'])) {
             $result = Application::$auth->proceedAuth($_POST['login'], $_POST['password']);
+
+            if($result &&
+                isset($_POST['user-remember']) && $_POST['user-remember'] == 'remember'){
+                $token = Application::$auth->generateToken();
+
+                UserRepository::setToken($_SESSION['auth']['id_user'], $token);
+            }
         }
         if (!$result) {
             return $this->render->renderPageWithForm([
@@ -169,15 +177,10 @@ class UserController extends Controller
         }
     }
 
-    #[NoReturn] public function actionLogout(): void
-    {
-
-        unset($_SESSION['user_authorized']);
-        unset($_SESSION['user_name']);
-        unset($_SESSION['id_user']);
-
+    #[NoReturn] public function actionLogout(): void {
+        UserRepository::destroyToken();
         session_destroy();
-
+        unset($_SESSION['auth']);
         header("Location: /");
         die();
     }
